Типы id и curid изменены на size_t

+++ libs/compiler/codegen.c
@@ -839,7 +839,7 @@ void output_export(universal_io *const io, const syntax *const sx)
 {
 	uni_printf(io, "#!/usr/bin/ruc-vm\n");
 
-	uni_printf(io, "%zi %i %i %zi %i %i %i\n", mem_get_size(sx), sx->funcnum, sx->id,
+	uni_printf(io, "%zi %i %zi %zi %i %i %i\n", mem_get_size(sx), sx->funcnum, sx->id,
 				   sx->rp, sx->md, sx->maxdisplg, sx->wasmain);
 
 	for (size_t i = 0; i < mem_get_size(sx); i++)
@@ -854,7 +854,7 @@ void output_export(universal_io *const io, const syntax *const sx)
 	}
 	uni_printf(io, "\n");
 
-	for (int i = 0; i < sx->id; i++)
+	for (size_t i = 0; i < sx->id; i++)
 	{
 		uni_printf(io, "%i ", sx->identab[i]);
 	}
